// src/generated/graphql.tsx
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** Parse string date to type Date */
  CacheDate: any;
  /** The javascript `Date` as string. Type represents date and time as the ISO Date string. */
  DateTime: any;
  /** The `JSONObject` scalar type represents JSON objects as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSONObject: any;
};
// ...
export type CitiesInput = {
  state_id: Scalars['String'];
};

export type City = {
  __typename?: 'City';
  id: Scalars['ID'];
  name: Scalars['String'];
  slug: Scalars['String'];
  status: Scalars['Boolean'];
  id_locale?: Maybe<Scalars['String']>;
  state: State;
};
// ...
export type Company = {
  __typename?: 'Company';
  id: Scalars['ID'];
  business_name: Scalars['String'];
  fantasy_name: Scalars['String'];
  email: Scalars['String'];
  phone_number: Scalars['String'];
  site: Scalars['String'];
  register_number: Scalars['String'];
  register_number_region: Scalars['Float'];
  address_one: Scalars['String'];
  address_two: Scalars['String'];
  zip_code: Scalars['String'];
  count_challenges: Scalars['Float'];
  challenges_subscription_count: Scalars['Float'];
  challenges_award_count: Scalars['Float'];
  company_type_id: Scalars['String'];
  city_id: Scalars['String'];
  active: Scalars['Boolean'];
  profile: Profile;
  company_addresses?: Maybe<Array<CompanyAddressType>>;
  created_at: Scalars['DateTime'];
  updated_at: Scalars['DateTime'];
};

export type CompanyAddressType = {
  __typename?: 'CompanyAddressType';
  address_id: Scalars['String'];
  company_id: Scalars['String'];
  is_billing_address: Scalars['Boolean'];
  address: AddressType;
  company: Company;
};
// ...
export type CreateCompanyInput = {
  business_name: Scalars['String'];
  fantasy_name?: Maybe<Scalars['String']>;
  email: Scalars['String'];
  phone_number: Scalars['String'];
  site?: Maybe<Scalars['String']>;
  register_number?: Maybe<Scalars['String']>;
  register_number_region?: Maybe<Scalars['Float']>;
  address_one?: Maybe<Scalars['String']>;
  address_two?: Maybe<Scalars['String']>;
  company_type_id?: Maybe<Scalars['String']>;
  zip_code?: Maybe<Scalars['String']>;
  city_id?: Maybe<Scalars['String']>;
};
// ...
export type Mutation = {
// ...
  createCompany: Company;
// ...
};
// ...
export type MutationCreateCompanyArgs = {
  data: CreateCompanyInput;
};
// ...
export type PaginationInput = {
  page: Scalars['Float'];
  offset: Scalars['Float'];
};
// ...
export type Profile = {
  __typename?: 'Profile';
  id: Scalars['ID'];
  username: Scalars['String'];
  followers_count: Scalars['Float'];
  following_count: Scalars['Float'];
  official: Scalars['Boolean'];
  profile_avatar?: Maybe<Scalars['String']>;
  profile_cover?: Maybe<Scalars['String']>;
  description?: Maybe<Scalars['String']>;
  user_id: Scalars['String'];
  company_id: Scalars['String'];
  is_follower?: Maybe<Scalars['Boolean']>;
  user: User;
};

export type ProfileSchedule = {
  __typename?: 'ProfileSchedule';
  _id: Scalars['ID'];
  profile_id: Scalars['String'];
  resource_key: Scalars['String'];
  date_occurrence?: Maybe<Scalars['DateTime']>;
  title: Scalars['String'];
  subtitle: Scalars['String'];
  deep_link: Scalars['String'];
  web_link: Scalars['String'];
  image_link: Scalars['String'];
};

export type Query = {
  __typename?: 'Query';
  getCity: Array<City>;
  getCityByCountry: Array<City>;
  getCitiesByName: Array<City>;
  getStates: Array<State>;
// ...
};


export type QueryGetCityArgs = {
  data: CitiesInput;
  pagination: PaginationInput;
};


export type QueryGetCityByCountryArgs = {
  country_id: Scalars['String'];
  pagination: PaginationInput;
};


export type QueryGetCitiesByNameArgs = {
  name: Scalars['String'];
  pagination: PaginationInput;
};


export type QueryGetStatesArgs = {
  pagination: PaginationInput;
};
// ...
export type State = {
  __typename?: 'State';
  id: Scalars['ID'];
  name: Scalars['String'];
  slug: Scalars['String'];
  abbreviation: Scalars['String'];
  active: Scalars['Boolean'];
  id_locale: Scalars['String'];
  country_id: Scalars['String'];
};